test: consolidate unit test suites
