+++ tests/SOURCE_GUARDS.md
@@ -11,7 +11,7 @@ extraction.
 | `tests/unit/auth/core.test.ts` | `auth-returnTo-threading` | Auth routing | Pure validator behavior is covered, but the redirect page/server data threading spans SvelteKit route files that are not all callable as isolated Bun units. | Add route/component harness coverage for the redirect load/page flow where onboarding context wins and safe `returnTo` is used otherwise. |
 | `tests/unit/security/basics.test.ts` | `hooks-anon-admin-returnTo` | Auth hooks | `authorizationHandle` is composed inside `hooks.server.ts` and not exported separately; source guard pins anonymous admin path encoding and non-admin dashboard redirect. | Export a behavior-neutral authorization helper or add hook-level request tests that can invoke this branch directly. |
 | `tests/unit/security/basics.test.ts` | `onboarding-bootstrap-origin` | Onboarding hooks | The bootstrap banner call is an initialization side effect inside the hook chain; the negative origin argument is not externally observable in Bun unit tests. | Extract the banner invocation decision or add a hook harness that spies on the banner call without importing request-derived origin. |
-| `tests/unit/onboarding/settings-substep-walk.test.ts` | `onboarding-settings-substeps` | Onboarding UI | The multi-substep footer/rendering behavior lives in a Svelte page and requires DOM interaction for true behavior coverage. | Replace with component/browser tests that walk the settings substeps and assert submit button gating. |
+| `tests/unit/onboarding/settings.test.ts` | `onboarding-settings-substeps` | Onboarding UI | The multi-substep footer/rendering behavior lives in a Svelte page and requires DOM interaction for true behavior coverage. | Replace with component/browser tests that walk the settings substeps and assert submit button gating. |
 | `tests/unit/plex-login.test.ts` | `plex-login-browser-safe-fields` | Auth client/API | Existing unit tests cover client helpers, but browser-safe response-shape constraints span auth endpoint and redirect load source without a direct exported schema. | Export a response DTO/schema or add route tests that assert completed auth payloads omit Plex tokens/private account fields. |
 | `tests/unit/security/basics.test.ts` | `admin-guard-hook-wiring` | Security/auth hooks | Guard helpers are tested behaviorally; the hook's use of the helper and redirect target are not separately exported. | Add hook-level tests or an exported authorization decision helper covering admin route probes. |
 | `tests/unit/security/dogfood-010-011.test.ts` | `plex-login-no-token-postMessage` | Security/auth client | This is a negative security invariant over absent popup/postMessage logging code; Bun unit tests cannot observe absent future branches without executing browser code. | Replace with browser/client tests that exercise the popup polling path and assert no token payload is posted or logged. |
