--- tests/unit/onboarding/page-action-claim-errors.test.ts
@@ -1,135 +1,0 @@
-import { beforeEach, describe, expect, it } from 'bun:test';
-import type { Cookies } from '@sveltejs/kit';
-import { db } from '$lib/server/db/client';
-import { appSettings } from '$lib/server/db/schema';
-import {
-	claimOnboardingInstance,
-	clearBootstrapToken,
-	createBootstrapToken,
-	ONBOARDING_CLAIM_REQUIRED_MESSAGE
-} from '$lib/server/onboarding/bootstrap';
-import { actions as completeActions } from '../../../src/routes/onboarding/complete/+page.server';
-import { actions as plexActions } from '../../../src/routes/onboarding/plex/+page.server';
-import { actions as settingsActions } from '../../../src/routes/onboarding/settings/+page.server';
-import { actions as syncActions } from '../../../src/routes/onboarding/sync/+page.server';
-
-type Action = (event: unknown) => Promise<unknown>;
-
-const adminLocals = {
-	user: { id: 1, plexId: 1, username: 'admin', isAdmin: true }
-} as unknown as App.Locals;
-
-const anonymousLocals = {} as App.Locals;
-
-const nonAdminLocals = {
-	user: { id: 2, plexId: 2, username: 'user', isAdmin: false }
-} as unknown as App.Locals;
-
-const claimCheckedActions = [
-	['plex.verifyAdmin', plexActions.verifyAdmin],
-	['plex.continueAfterServerSelection', plexActions.continueAfterServerSelection],
-	['plex.forceManualSelection', plexActions.forceManualSelection],
-	['plex.confirmOwnershipOverride', plexActions.confirmOwnershipOverride],
-	['settings.saveSettings', settingsActions.saveSettings],
-	['settings.skipSettings', settingsActions.skipSettings],
-	['settings.testAIConnection', settingsActions.testAIConnection],
-	['sync.startSync', syncActions.startSync],
-	['sync.cancelSync', syncActions.cancelSync],
-	['sync.continue', syncActions.continue],
-	['complete.goToDashboard', completeActions.goToDashboard]
-] as const;
-
-const adminRequiredActions = [
-	['settings.saveSettings', settingsActions.saveSettings],
-	['settings.skipSettings', settingsActions.skipSettings],
-	['settings.testAIConnection', settingsActions.testAIConnection],
-	['sync.startSync', syncActions.startSync],
-	['sync.cancelSync', syncActions.cancelSync],
-	['sync.continue', syncActions.continue],
-	['complete.goToDashboard', completeActions.goToDashboard]
-] as const;
-
-function makeCookies(errorToThrow?: Error): Cookies {
-	const values = new Map<string, string>();
-	return {
-		get: () => {
-			if (errorToThrow) throw errorToThrow;
-			return values.get('obzorarr_onboarding_claim');
-		},
-		getAll: () => [],
-		set: (name: string, value: string) => {
-			values.set(name, value);
-		},
-		delete: (name: string) => {
-			values.delete(name);
-		},
-		serialize: () => ''
-	} as unknown as Cookies;
-}
-
-async function runAction(
-	action: unknown,
-	cookies: Cookies,
-	locals: App.Locals = adminLocals
-): Promise<unknown> {
-	const request = new Request('http://localhost/onboarding', {
-		method: 'POST',
-		body: new FormData()
-	});
-
-	return (action as Action)({
-		request,
-		locals,
-		cookies,
-		url: new URL(request.url)
-	});
-}
-
-describe('onboarding page action claim checks', () => {
-	beforeEach(async () => {
-		await db.delete(appSettings);
-		clearBootstrapToken();
-	});
-
-	for (const [name, action] of claimCheckedActions) {
-		it(`${name} returns the expected claim-required failure`, async () => {
-			const result = await runAction(action, makeCookies());
-
-			expect(result).toMatchObject({
-				status: 403,
-				data: { error: ONBOARDING_CLAIM_REQUIRED_MESSAGE }
-			});
-		});
-
-		it(`${name} propagates unexpected claim errors`, async () => {
-			const unexpected = new Error(`unexpected claim renewal failure in ${name}`);
-
-			try {
-				await runAction(action, makeCookies(unexpected));
-				expect.unreachable('Expected unexpected claim error to be thrown');
-			} catch (err) {
-				expect(err).toBe(unexpected);
-			}
-		});
-	}
-
-	for (const [name, action] of adminRequiredActions) {
-		it(`${name} preserves the admin failure after a valid setup claim`, async () => {
-			const cookies = makeCookies();
-			const token = createBootstrapToken();
-			expect(await claimOnboardingInstance(cookies, token)).toBe('claimed');
-
-			const anonymousResult = await runAction(action, cookies, anonymousLocals);
-			expect(anonymousResult).toMatchObject({
-				status: 403,
-				data: { error: 'Admin access required' }
-			});
-
-			const nonAdminResult = await runAction(action, cookies, nonAdminLocals);
-			expect(nonAdminResult).toMatchObject({
-				status: 403,
-				data: { error: 'Admin access required' }
-			});
-		});
-	}
-});
